Change search field type to "search"
This is like "text" except that a little "x" button appears so that one can
easily clear the field. Also, on mobile browsers, the action button on the
keyboard will display "Search" instead of just "Go".

Old browsers will gracefully fallback to "text" anyways, so this is a
progressive enchancement.